When downloading files, the name of the zip will change accordingly
If downloading a single folder, the name of the downloaded zip will be the name of the folder (if it's a folder).
If downloading multiple items, the name will be a date timestamp.
If none of the criteria matches above, it will default to files.zip
When downloading all files from a document space, the name will use the name of the document space